feat: Propertyインタフェースを導入し、プロパティ関係についてはこれを継承するように実装する。
