+++ src/main/java/feature/property/Property.java
@@ -1,0 +1,14 @@
+package feature.property;
+
+/**
+ * <p> プロパティインタフェース </p>
+ *
+ * <p>
+ *     プロパティを定義するインタフェースです。
+ * </p>
+ */
+public interface Property {
+
+    @Override
+    String toString();
+}
